+++ src/rgbdmux.rs
@@ -103,6 +103,11 @@ impl ObjectImpl for RgbdMux {
 }
 
 impl ElementImpl for RgbdMux {
+    /// This function provides a custom implementation to what should happen when request pads are
+    /// released.
+    /// # Arguments
+    /// * `element` - The element that represents `rgbdmux` in GStreamer.
+    /// * `pad` - The pad that is soon to be released.
     fn release_pad(&self, element: &gst::Element, pad: &gst::Pad) {
         // De-activate the pad
         pad.set_active(false)
@@ -115,10 +120,14 @@ impl ElementImpl for RgbdMux {
 
         // Remove the pad from our internal reference HashMap
         let pad_name = pad.get_name().as_str().to_string();
+        gst_debug!(self.cat, obj: element, "release_pad: {}", pad_name);
         self.sink_pads
             .lock()
             .expect("Could not lock sink_pads")
             .retain(|x| *x != pad_name);
+
+        // TODO: We should check whether we're in the process of shutting down before calling this
+        self.renegotiate_downstream_caps(element);
     }
 }
 
@@ -166,6 +175,13 @@ impl AggregatorImpl for RgbdMux {
         self.finish_buffer(aggregator, main_output_buffer)
     }
 
+    /// This function is called when a peer element requests a pad. It provides a custom implementation
+    /// for how the pad should be created.
+    /// # Arguments
+    /// * `aggregator` - The element that represents the `rgbdmux` in GStreamer.
+    /// * `_templ` - (not used) The template that should be used in pad creation.
+    /// * `req_name` - The requested name for the pad.
+    /// * `_caps` - (not used) The CAPS to use for the pad.
     fn create_new_pad(
         &self,
         aggregator: &gst_base::Aggregator,
@@ -174,12 +190,14 @@ impl AggregatorImpl for RgbdMux {
         _caps: Option<&gst::Caps>,
     ) -> Option<gst_base::AggregatorPad> {
         match req_name {
-            // Make sure the name is valid
-            None => {
-                gst_error!(self.cat, obj: aggregator, "Invalid request pad name");
-                return None;
-            }
-            Some(name) => {
+            Some(name) if name.starts_with("sink_") => {
+                let mut sink_pads = self.sink_pads.lock().expect("Cannot lock sink_pads");
+                gst_debug!(
+                    self.cat,
+                    obj: aggregator,
+                    "create_new_pad for name: {}",
+                    name
+                );
                 // Create new sink pad from the template
                 let new_sink_pad = gst::Pad::new_from_template(
                     &aggregator
@@ -191,12 +209,7 @@ impl AggregatorImpl for RgbdMux {
                 .expect("Could not cast pad to AggregatorPad");
 
                 // Drop all buffers on already existing pads (if any)
-                for pad_name in self
-                    .sink_pads
-                    .lock()
-                    .expect("Could nok lock sink_pads")
-                    .iter()
-                {
+                for pad_name in sink_pads.iter() {
                     loop {
                         if aggregator
                             .get_static_pad(pad_name)
@@ -213,13 +226,8 @@ impl AggregatorImpl for RgbdMux {
                     }
                 }
 
-                // TODO (important): fix downstream re-negotiation
-
                 // Insert the new sink pad name into the struct
-                self.sink_pads
-                    .lock()
-                    .expect("Could not lock sink_pads")
-                    .push(name.to_string());
+                sink_pads.push(name.to_string());
 
                 // Activate the sink pad
                 new_sink_pad
@@ -228,7 +236,53 @@ impl AggregatorImpl for RgbdMux {
 
                 Some(new_sink_pad)
             }
+            _ => {
+                gst_error!(
+                    self.cat,
+                    obj: aggregator,
+                    "Invalid request pad name. Only sink pads may be requested."
+                );
+                return None;
+            }
         }
+    }
+
+    /// This function is called during CAPS negotiation. It can be used to decide on a CAPS format
+    /// or delay the negotiation until sufficient data is present to decide on the CAPS (in this
+    /// case when an upstream element has requested sink pads)
+    /// # Arguments
+    /// * `aggregator` - A reference to the element that represents `rgbdmux` in GStreamer.
+    ///* `_caps` - (not used) The CAPS that is currently negotiated for the element.
+    fn update_src_caps(
+        &self,
+        aggregator: &gst_base::Aggregator,
+        _caps: &gst::Caps,
+    ) -> Result<gst::Caps, gst::FlowError> {
+        gst_debug!(self.cat, "update_src_caps");
+        // Check how many sink pads has been created
+        let no_sink_pads = {
+            self.sink_pads
+                .lock()
+                .expect("Could not lock sink pads")
+                .len()
+        };
+        // if no sink pads are present, we're not ready to negotiate CAPS, otherwise do the negotiation
+        match no_sink_pads {
+            0 => Err(gst_base::AGGREGATOR_FLOW_NEED_DATA), // we're not ready to decide on CAPS yet
+            _ => Ok(self.get_current_downstream_caps(aggregator.upcast_ref::<gst::Element>())),
+        }
+    }
+
+    /// This function is used to hint the type of CAPS we're expecting in the element.
+    /// Right now the function simply ignores the suggested CAPS and pushes the one generated from
+    /// the video/rgbd CAPS.
+    /// # Arguments
+    /// * `aggregator` - A reference to the element that represents `rgbdmux` in GStreamer.
+    ///* `_caps` - (not used) The CAPS that is currently negotiated for the element.
+    fn fixate_src_caps(&self, aggregator: &gst_base::Aggregator, _caps: gst::Caps) -> gst::Caps {
+        let elm = aggregator.upcast_ref::<gst::Element>();
+        gst_debug!(self.cat, obj: elm, "fixate_src_caps");
+        self.get_current_downstream_caps(elm)
     }
 }
 
@@ -256,8 +310,126 @@ impl RgbdMux {
         tags.get_mut()
             .ok_or(MuxingError("Could not get mutable reference to tags"))?
             .add::<gst::tags::Title>(&&pad_name[5..], gst::TagMergeMode::Append);
-        TagsMeta::add(buffer.make_mut(), &mut tags);
+        let mut_buffer = buffer.get_mut().ok_or(MuxingError("Could not get mutable reference to buffer"))?;
+        TagsMeta::add(mut_buffer, &mut tags);
         Ok(buffer)
+    }
+
+    /// Extracts the relevant fields from the pad's CAPS and converts them into a tuple containing
+    /// the field's name as the first and its value as second.
+    /// # Arguments
+    /// * `pad_caps` - A reference to the pad's CAPS.
+    /// * `pad_name` - The name of the stream we're currently generating CAPS for.
+    fn push_sink_caps_format(
+        &self,
+        pad_caps: &gst::Caps,
+        pad_name: &str,
+        src_caps: &mut gst::StructureRef,
+    ) {
+        let pad_caps = pad_caps.iter().next().expect("Got empty CAPS in rgbdmux");
+        let stream_name = &pad_name[5..];
+
+        // Filter out all CAPS we don't care about and map those we do into strings
+        for (field, value) in pad_caps.iter() {
+            match field {
+                "format" => {
+                    let src_field_name = format!("{}_{}", stream_name, field);
+                    src_caps.set(&src_field_name, &value.get::<&str>().unwrap());
+                }
+                "width" => {
+                    let src_field_name = format!("{}_{}", stream_name, field);
+                    src_caps.set(&src_field_name, &value.get::<i32>().unwrap());
+                }
+                "height" => {
+                    let src_field_name = format!("{}_{}", stream_name, field);
+                    src_caps.set(&src_field_name, &value.get::<i32>().unwrap());
+                }
+                _ => {
+                    gst_info!(
+                        self.cat,
+                        "Ignored CAPS field {} of stream {}",
+                        field,
+                        stream_name,
+                    );
+                }
+            }
+        }
+    }
+
+    /// Get the current downstream CAPS. The downstream CAPS are generated based on the current sink
+    /// pads on the muxer.
+    /// # Arguments
+    /// * `element` - The element that represents the `rgbdmux`.
+    fn get_current_downstream_caps(&self, element: &gst::Element) -> gst::Caps {
+        // First lock sink_pads, so that we may iterate it
+        let sink_pads = self
+            .sink_pads
+            .lock()
+            .expect("Failed to obtain `sink_pads` lock.");
+
+        // Join all the pad names to create the 'streams' section of the CAPS
+        let streams = sink_pads
+            .iter()
+            .map(|s| &s[5..])
+            .collect::<Vec<&str>>()
+            .join(",");
+
+        // TODO: Remove hardcoded framerate
+        let mut downstream_caps = gst::Caps::new_simple(
+            "video/rgbd",
+            &[
+                ("streams", &streams),
+                ("framerate", &gst::Fraction::new(30, 1)),
+            ],
+        );
+        let mut_caps = downstream_caps
+            .make_mut()
+            .get_mut_structure(0)
+            .expect("Could not get mutable CAPS in rgbdmux");
+
+        // Map the caps into their corresponding stream formats
+        for pad_name in sink_pads.iter() {
+            // First find the current CAPS of Pad we're currently dealing with
+            let pad_caps = element
+                .get_static_pad(pad_name)
+                .expect(&format!(
+                    "Could not get static pad from aggregator with name `{}`",
+                    pad_name
+                ))
+                .get_current_caps();
+            match pad_caps {
+                Some(pc) => self.push_sink_caps_format(&pc, pad_name, mut_caps),
+                None => { /*ignore*/ }
+            }
+        }
+
+        gst_info!(
+            self.cat,
+            obj: element,
+            "stream_caps were found to be: {:?}.",
+            downstream_caps
+        );
+
+        downstream_caps.to_owned()
+    }
+
+    /// Generates and sends new CAPS for the downstream elements. This function automatically
+    /// generates CAPS based on the sink_pads of the muxer and their current CAPS.
+    /// # Arguments
+    /// * `element` - The element that represents the `rgbdmux`.
+    fn renegotiate_downstream_caps(&self, element: &gst::Element) {
+        gst_debug!(self.cat, obj: element, "renegotiate_downstream_caps");
+        // Figure out the new caps the element should output
+        let ds_caps = self.get_current_downstream_caps(element);
+        // And send a CAPS event downstream
+        let caps_event = gst::Event::new_caps(&ds_caps).build();
+        if !element.send_event(caps_event) {
+            gst_error!(
+                self.cat,
+                obj: element,
+                "Failed to send CAPS negotiation event"
+            );
+        }
     }
 }